config/script_bootstrap_db/script/requirements.txt
Faker==37.3.0
mysql-connector-python==9.3.0
